[TM-1842] Ditch the "HasLinkedFields" trait.
This has been a common source of confusion. It was included on a bunch of models that weren't actually using it. The only place it _was_ used is on FormQuestion, so I've simply moved the methods directly to that model.
