+++ app/Models/V2/Forms/FormQuestion.php
@@ -3,15 +3,17 @@
 namespace App\Models\V2\Forms;
 
 use App\Models\Traits\HasI18nTranslations;
-use App\Models\Traits\HasLinkedFields;
 use App\Models\Traits\HasUuid;
 use App\Models\V2\I18n\I18nItem;
+use App\Models\V2\Organisation;
+use App\Models\V2\ProjectPitch;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Arr;
 
 class FormQuestion extends Model
 {
@@ -19,7 +21,6 @@ class FormQuestion extends Model
     use SoftDeletes;
     use HasUuid;
     use HasI18nTranslations;
-    use HasLinkedFields;
 
     protected $fillable = [
         'form_section_id',
@@ -126,5 +127,74 @@ class FormQuestion extends Model
     public function getRouteKeyName()
     {
         return 'uuid';
+    }
+
+    public function getLinkedFieldInfo(array $params = null): ?array
+    {
+        $info = $this->getLinkedFieldConfig();
+
+        if (is_array($info)) {
+            $info['uuid'] = $this->getLinkedModelUuid($info, $params);
+        }
+
+        return $info;
+    }
+
+    protected function getLinkedFieldConfig(): ?array
+    {
+        $types = ['fields', 'file-collections', 'relations'];
+        foreach (config('wri.linked-fields.models', []) as $modelKey => $cfgModel) {
+            foreach ($types as $type) {
+                if (Arr::has($cfgModel, $type. '.' . $this->linked_field_key)) {
+                    $questionProps = data_get($cfgModel[$type], $this->linked_field_key);
+                    $modelProps = [
+                        'model-label' => data_get($cfgModel, 'label'),
+                        'model-key' => $modelKey,
+                        'model' => data_get($cfgModel, 'model'),
+                        'link-type' => $type,
+                    ];
+
+                    return array_merge($modelProps, $questionProps);
+                }
+            }
+        }
+
+        return null;
+    }
+
+    protected function getLinkedModelUuid($cfg, array $params = null): ?string
+    {
+        if (! empty(data_get($params, 'model_uuid'))) {
+            return data_get($params, 'model_uuid');
+        }
+
+        return $this->getApplicationModelsUuids($cfg, $params);
+    }
+
+    protected function getApplicationModelsUuids($cfg, array $params = null): ?string
+    {
+        $organisation = data_get($params, 'organisation', null);
+        if (empty($organisation) && ! empty(data_get($params, 'organisation_uuid', null))) {
+            $organisation = Organisation::isUuid(data_get($params, 'organisation_uuid'))->first();
+        }
+
+        $projectPitch = data_get($params, 'project-pitch', null);
+
+        if (empty($projectPitch) && ! empty(data_get($params, 'project_pitch_uuid', null))) {
+            $projectPitch = ProjectPitch::isUuid(data_get($params, 'project_pitch_uuid'))->first();
+        }
+
+        if ($organisation && empty($projectPitch)) {
+            $projectPitch = $organisation->projectPitches()->first();
+        }
+
+        switch ($cfg['model-key']) {
+            case 'organisation':
+                return empty($organisation) ? null : $organisation->uuid;
+            case 'project-pitch':
+                return empty($projectPitch) ? null : $projectPitch->uuid;
+            default:
+                return null;
+        }
     }
 }
