+++ app/Models/V2/Nurseries/Nursery.php
@@ -6,7 +6,6 @@ use App\Models\Framework;
 use App\Models\Traits\HasEntityResources;
 use App\Models\Traits\HasEntityStatus;
 use App\Models\Traits\HasFrameworkKey;
-use App\Models\Traits\HasLinkedFields;
 use App\Models\Traits\HasUpdateRequests;
 use App\Models\Traits\HasUuid;
 use App\Models\Traits\HasV2MediaCollections;
@@ -37,7 +36,6 @@ class Nursery extends Model implements MediaModel, AuditableContract, EntityMode
     use HasFactory;
     use HasUuid;
     use SoftDeletes;
-    use HasLinkedFields;
     use UsesLinkedFields;
     use InteractsWithMedia;
     use HasV2MediaCollections;
