+++ app/Models/V2/Nurseries/NurseryReport.php
@@ -5,7 +5,6 @@ namespace App\Models\V2\Nurseries;
 use App\Models\Framework;
 use App\Models\Traits\HasEntityResources;
 use App\Models\Traits\HasFrameworkKey;
-use App\Models\Traits\HasLinkedFields;
 use App\Models\Traits\HasReportStatus;
 use App\Models\Traits\HasUpdateRequests;
 use App\Models\Traits\HasUuid;
@@ -41,7 +40,6 @@ class NurseryReport extends Model implements MediaModel, AuditableContract, Repo
     use HasUuid;
     use SoftDeletes;
     use HasReportStatus;
-    use HasLinkedFields;
     use UsesLinkedFields;
     use InteractsWithMedia;
     use HasV2MediaCollections;
